rename classes to external html
