+++ mzmine-community/src/main/java/io/github/mzmine/modules/visualization/external_row_html/ExternalRowHtmlVisualizerController.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright (c) 2004-2025 The mzmine Development Team
+ *
+ * Permission is hereby granted, free of charge, to any person
+ * obtaining a copy of this software and associated documentation
+ * files (the "Software"), to deal in the Software without
+ * restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following
+ * conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+ * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+ * OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package io.github.mzmine.modules.visualization.external_row_html;
+
+import io.github.mzmine.datamodel.features.FeatureListRow;
+import io.github.mzmine.gui.framework.fx.FeatureRowInterfaceFx;
+import io.github.mzmine.javafx.mvci.FxController;
+import io.github.mzmine.javafx.mvci.FxViewBuilder;
+import java.util.List;
+import org.jetbrains.annotations.NotNull;
+
+public class ExternalRowHtmlVisualizerController extends
+    FxController<ExternalRowHtmlVisualizerModel> implements FeatureRowInterfaceFx {
+
+  private final ExternalRowHtmlVisualizerViewBuilder viewBuilder;
+
+  public ExternalRowHtmlVisualizerController() {
+    super(new ExternalRowHtmlVisualizerModel());
+    viewBuilder = new ExternalRowHtmlVisualizerViewBuilder(model);
+  }
+
+  @Override
+  protected @NotNull FxViewBuilder<ExternalRowHtmlVisualizerModel> getViewBuilder() {
+    return viewBuilder;
+  }
+
+  @Override
+  public boolean hasContent() {
+    return true;
+  }
+
+  @Override
+  public void setFeatureRows(final @NotNull List<? extends FeatureListRow> selectedRows) {
+    model.setSelectedRow(selectedRows.isEmpty() ? null : selectedRows.getFirst());
+  }
+}
