--- mzmine-community/src/main/java/io/github/mzmine/modules/visualization/masst/MasstVisualizerTab.java
@@ -1,13 +1,0 @@
-package io.github.mzmine.modules.visualization.masst;
-
-import io.github.mzmine.gui.mainwindow.SimpleTab;
-
-public class MasstVisualizerTab extends SimpleTab {
-
-  public MasstVisualizerTab() {
-    super("MASST");
-
-    MasstVisualizerController controller = new MasstVisualizerController();
-    setContent(controller.getViewBuilder().build());
-  }
-}
